rename file to correct spelling
